Unit32 timer, and moved check to CLControlStep
