+++ firmware/src/common.h
@@ -121,8 +121,7 @@ typedef enum
 	ERROR_ENCODER_READING_UNSTABLE = 9,
 	ERROR_PLANNER_INVALID_INPUT = 10,
 	ERROR_PLANNER_VCRUISE_OVER_LIMIT = 11,
-	ERROR_HALL_SENSOR_CALIBRATION_FAILED = 12,
-	ERROR_WATCHDOG_TIMEOUT = 13
+	ERROR_HALL_SENSOR_CALIBRATION_FAILED = 12
 } Error;
 
 #endif // #ifndef COMMON_H
