+++ firmware/src/watchdog/watchdog.c
@@ -1,8 +1,8 @@
-#include "src/can/watchdog.h"
+#include "src/watchdog/watchdog.h"
 #include "src/common.h"
 
 static TimeoutWatchdog watchdog = {
-    .timeout_cycles = 10000,
+    .timeout_cycles = PWM_FREQ_HZ, // Set default 1s timeout
     .counter = 0,
     .enabled = false
 };
@@ -27,14 +27,15 @@ PAC5XXX_RAMFUNC bool Watchdog_increment_and_check(void)
     return false;
 }
 
-void Watchdog_set_timeout_cycles(uint16_t cycles)
+void Watchdog_set_timeout_cycles(uint32_t cycles)
 {
     watchdog.timeout_cycles = cycles;
 }
 
 void Watchdog_set_timeout_s(float s)
 {
-    uint16_t cycles = (uint16_t)(s / PWM_PERIOD_S);
+    // Maximum 2^32/20000 = 214748 seconds
+    uint32_t cycles = (uint32_t)(s * PWM_FREQ_HZ);
     Watchdog_set_timeout_cycles(cycles);
 }
 
@@ -43,6 +44,7 @@ void Watchdog_enable(void)
     Watchdog_reset();
     watchdog.enabled = true;
 }
+
 void Watchdog_disable(void)
 {
     watchdog.enabled = false;
